backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
sqlalchemy==2.0.23
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
cryptography>=3.4.8
pycryptodome>=3.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1